+++ src/index.js
@@ -28,4 +28,4 @@ Asker.prototype.batch = batch
 
 export default Asker
 export { splitBlob, safeCall } from './public'
-export { object2Query } from './util/format'
+export { object2Query, query2Object } from './util/format'